Fix false positive in `Lint/AssignmentInCondition` (#3520) (#3529)
This was broken in b0d8e0cf which was simply a refactoring commit. It looks like part of the code was omitted accidentally when porting to the new block. Specifically, the part which was dropped was: https://github.com/bbatsov/rubocop/blob/78de463f33992bce5b1ecf40e311958ad773bad5/lib/rubocop/cop/lint/assignment_in_condition.rb#L35-L38

I've ported the missing checks to the `#skip_children?` method, and added a test which would've caught the regression.