Allow `#to_yml` to produce single-quoted strings
In Ruby 2.0.0, the YAML generated for a single cop that has a partial
`StyleGuide` URL fragment has that fragment delimited by single
quotes, whereas in all other Ruby 2.x builds it is delimited by double
quotes. This was not previously an issue because the full URLs that
were used for `StyleGuide` settings are not delimited by anything.

This is very much an imperfect fix -- and as soon as it is decided
that RuboCop no longer needs to support 2.0.0 it should be reversed.
Still, it's preferable to marking each of these two failing specs as
broken in that Ruby build.