[Fix #3577] Fix `Style/RaiseArgs` not allowing raise with splatted args
With the default style (exploded) one is typically allowed to use the compact style if passing multiple args into the constructor. However an offence was being generated if those args were splatted in, this changes that so there is no offence.

Also refactored the cop to use guard clauses to reduce nesting.